Extension works now
The extension is now in a working state, and correctly communicates
requests from the hijacked fetch() to the background, and then to
innertube_proxy server. Responses are correctly handled in the
opposite direction as well. There are some more feature additions
to be added in later, but for now the extension is usable and I can
focus on adding functionality to the server.